FLOE-529: refactored index.js, split into src/js folder
