+++ src/js/dataSource.js
@@ -12,8 +12,6 @@ https://raw.githubusercontent.com/waharnum/sjrk-storyTelling/master/LICENSE.txt
 var fluid = require("infusion");
 require("kettle");
 
-// var sjrk = fluid.registerNamespace("sjrk");
-
 fluid.defaults("sjrk.storyTelling.server.dataSource", {
     gradeNames: ["kettle.dataSource.URL", "kettle.dataSource.CouchDB"],
     url: "http://localhost:5984/stories/%storyId",
@@ -22,14 +20,3 @@ fluid.defaults("sjrk.storyTelling.server.dataSource", {
     },
     writable: true
 });
-
-// Writable usage example
-// var dataSource = sjrk.storyTelling.server.dataSource();
-//
-// var promise = dataSource.set({directStoryId: "588"}, {"title": "Story", "author": "Alan", "content": "My content"});
-//
-// promise.then(function (response) {
-//     console.log("Got dataSource response of ", response);
-// }, function (error) {
-//     console.error("Got dataSource error response of ", error);
-// });
